digital-credentials/concurrent-requests.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Digital Credential concurrent request rejection tests</title>
<link rel="help" href="https://wicg.github.io/digital-credentials/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body></body>
<script type="module">
  import { makeGetOptions } from "./support/helper.js";

  promise_test(async (t) => {
    const abortController = new AbortController();
    const firstOptions = makeGetOptions({ signal: abortController.signal });
    const secondOptions = makeGetOptions();

    await test_driver.bless("user activation for first request");
    const firstRequest = navigator.credentials.get(firstOptions);

    // Fresh user activation ensures any rejection is due to the pending
    // request, not lack of user activation (both would be NotAllowedError)
    await test_driver.bless("user activation for second request");
    const secondRequest = navigator.credentials.get(secondOptions);

    await promise_rejects_dom(
      t,
      "NotAllowedError",
      secondRequest,
      "Second concurrent request should be rejected with NotAllowedError"
    );

    abortController.abort();
    await promise_rejects_dom(
      t,
      "AbortError",
      firstRequest,
      "First request should be aborted"
    );
  }, "A second get() call while another is pending should reject with NotAllowedError.");
</script>
